Use terraform to create roles for main module
To insist on using least-privileged roles from the very
beginning, terraform itself is used to create roles to be
used by terraform. The code under `terraform-role` should be
applied as a role / user with privileges in IAM role creation.
It will provision a role that'll have just enough privileges
to perform all the actions in the main repo.